Updated the build system

+++ scripts/Clean.hx
@@ -3,7 +3,11 @@ using Lambda;
 
 /** Runs the script. **/
 function main() {
-	["bin/mc2it_theme.js", "lib/index.js"].filter(FileSystem.exists).iter(FileSystem.deleteFile);
+	for (file in ["bin/mc2it_theme.js", "lib/index.js"]) {
+		if (FileSystem.exists(file)) FileSystem.deleteFile(file);
+		if (FileSystem.exists('$file.map')) FileSystem.deleteFile('$file.map');
+	}
+
 	Tools.cleanDirectory("var");
-	Tools.removeDirectory("www/css");
+	if (FileSystem.exists("www/css")) Tools.removeDirectory("www/css");
 }
